+++ src/ParquetColumn.cpp
@@ -93,7 +93,7 @@ uint32_t ParquetColumn::getValueSize() {
 }
 
 
-uint64_t ParquetColumn::getValues(uint8_t* vector, uint64_t num, uint8_t* nullvector, uint32_t* fkvector, uint32_t& fk) {
+uint64_t ParquetColumn::getValues(uint8_t* vector, uint64_t num, uint8_t* nullvector, uint64_t* fkvector, uint64_t& fk) {
 	if (cur_page == nullptr) return 0;
 	uint64_t count = cur_page->getValues(vector, num, nullvector, fkvector, fk);
 	while (count < num && cur_page != nullptr) {
